+++ Library_Management_System/src/main/java/com/library/Library_Management_System/Services/BookService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface BookService {
     Book bookSaver(Book book);
     List<Book> getAllBooks();
-    Book getBookById(int id);
-    void deleteById(int id);
+    Book getBookById(Long id);
+    void deleteById(Long id);
 }
